+++ src/collective/iconifiedcategory/browser/contentview.py
@@ -24,6 +24,8 @@ class FormMixin(object):
         'to_print': 'to_be_printed_activated',
         'to_sign': 'signed_activated',
         'signed': 'signed_activated',
+        'to_approve': 'approved_activated',
+        'approved': 'approved_activated',
     }
 
     @property
@@ -64,6 +66,8 @@ class BaseView(DefaultView):
         'to_print': 'to_be_printed_activated',
         'to_sign': 'signed_activated',
         'signed': 'signed_activated',
+        'to_approve': 'approved_activated',
+        'approved': 'approved_activated',
     }
 
     def updateWidgets(self):