+++ src/collective/iconifiedcategory/utils.py
@@ -465,6 +465,23 @@ def signed_message(obj=None, to_sign_value=None, signed_value=None):
         return messages[signed_value]
 
 
+def approved_message(obj=None, to_approve_value=None, approved_value=None):
+    """Return the approved message for the given object"""
+    messages = {
+        False: u'Element must be approved but is still not',
+        True: u'Element is approved',
+    }
+    not_to_approve_msg = u'Element should not be approved'
+    if obj:
+        if getattr(obj, 'to_approve', False) is False:
+            return not_to_approve_msg
+        return messages.get(getattr(obj, 'approved', False), '')
+    elif to_approve_value is False:
+        return not_to_approve_msg
+    else:
+        return messages[approved_value]
+
+
 def boolean_message(obj=None, attr_name='', value=None):
     """Return the default boolean status message for the given object"""
     messages = {